Refactoring to allow test cases to be created. The previous classes were not testable.

+++ samples/java/sbt.sample.app/src/com/ibm/sbt/sample/app/GetAllCommunitiesApp.java
@@ -22,8 +22,10 @@ import com.ibm.commons.runtime.Application;
 import com.ibm.commons.runtime.Context;
 import com.ibm.commons.runtime.RuntimeFactory;
 import com.ibm.commons.runtime.impl.app.RuntimeFactoryStandalone;
+import com.ibm.sbt.security.authentication.AuthenticationException;
 import com.ibm.sbt.services.client.connections.communities.Community;
 import com.ibm.sbt.services.client.connections.communities.CommunityService;
+import com.ibm.sbt.services.client.connections.communities.CommunityServiceException;
 import com.ibm.sbt.services.client.connections.communities.Member;
 import com.ibm.sbt.services.endpoints.BasicEndpoint;
 import com.ibm.sbt.services.endpoints.EndpointFactory;
@@ -34,33 +36,60 @@ import com.ibm.sbt.services.endpoints.EndpointFactory;
  */
 public class GetAllCommunitiesApp {
 
+    private RuntimeFactory runtimeFactory;
+    private Context context;
+    private Application application;
+    private BasicEndpoint endpoint;
+    private CommunityService communityService;
+    
+    public void init() throws AuthenticationException{
+        runtimeFactory = new RuntimeFactoryStandalone();
+        application = runtimeFactory.initApplication(null);
+        context = Context.init(application, null, null);
+        communityService = new CommunityService();
+        endpoint = (BasicEndpoint)EndpointFactory.getEndpoint(CommunityService.DEFAULT_ENDPOINT_NAME);
+        endpoint.login("admin", "***REMOVED***", true);
+    }
+    
+    public void destroy(){
+        if (context != null)
+            Context.destroy(context);
+        if (application != null)
+            Application.destroy(application);
+    }
+    
+    public Collection<Community> getPublicCommunities() throws CommunityServiceException{
+        return communityService.getPublicCommunities();
+    }
+    
+    public Member[] getCommunityMembers(Community community) throws CommunityServiceException{
+        return communityService.getMembers(community);
+    }
+    
 	/**
 	 * @param args
 	 */
 	public static void main(String[] args) {
-		RuntimeFactory runtimeFactory = new RuntimeFactoryStandalone();
-		Application application = runtimeFactory.initApplication(null);
-		Context context = Context.init(application, null, null);
-
-		try {
-			BasicEndpoint endpoint = (BasicEndpoint)EndpointFactory.getEndpoint(CommunityService.DEFAULT_ENDPOINT_NAME);
-	        endpoint.login("admin", "***REMOVED***", true);
-			
-	        CommunityService communityService = new CommunityService();
-            Collection<Community> communities = communityService.getPublicCommunities();
+	    GetAllCommunitiesApp app = new GetAllCommunitiesApp();
+		
+	    try {
+            app.init();
+            Collection<Community> communities = app.getPublicCommunities();
             for (Iterator<Community> iter = communities.iterator(); iter.hasNext(); ) {
                 Community community = iter.next();
-                Member[] members = communityService.getMembers(community);
+                Member[] members = app.getCommunityMembers(community);
                 System.out.println(community.getTitle());
                 for (int i=0; i<members.length; i++) {
                     System.out.println("    " + members[i].getEmail());
                 }
             }
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-
-		Context.destroy(context);
-		Application.destroy(application);
+        } catch (AuthenticationException e) {
+            e.printStackTrace();
+        } catch (CommunityServiceException e) {
+            e.printStackTrace();
+        } finally {
+            app.destroy();
+        }
+	    
 	}
 }
